requirements.txt
aiogram
python-dotenv
DrissionPage
prometheus_flask_exporter
flask
google-auth
google-auth-oauthlib
google-api-python-client
requests
